♻️ Refactor cloud servers creation
