+++ tests/Unit/Services/Cloud/DigitalOceanDropletCreatorTest.php
@@ -1,12 +1,6 @@
 <?php
 
-use App\Services\DigitalOcean\CreateDigitalOceanDropletData;
-use App\Services\DigitalOcean\DigitalOceanDropletCreator;
-use GuzzleHttp\Psr7\HttpFactory;
-use GuzzleHttp\Psr7\Response;
-use Psr\Http\Client\ClientInterface;
-use Psr\Http\Message\RequestInterface;
-use Psr\Http\Message\ResponseInterface;
+use App\Services\Cloud\DigitalOcean\CreateDigitalOceanDropletData;use App\Services\Cloud\DigitalOcean\DigitalOceanDropletCreator;use GuzzleHttp\Psr7\HttpFactory;use GuzzleHttp\Psr7\Response;use Psr\Http\Client\ClientInterface;use Psr\Http\Message\RequestInterface;use Psr\Http\Message\ResponseInterface;
 
 test('DigitalOceanDropletCreator uploads a raw SSH key and returns droplet details', function () {
     $client = new class([new Response(201, [], json_encode(['ssh_key' => ['id' => 44, 'name' => 'uploaded-key']])), new Response(202, [], json_encode(['droplet' => ['id' => 999, 'name' => 'sword-web', 'region' => ['slug' => 'ams3'], 'size_slug' => 's-1vcpu-1gb', 'status' => 'new']])), new Response(200, [], json_encode(['droplet' => ['networks' => ['v4' => [['type' => 'public', 'ip_address' => '203.0.113.10']]]]]))]) implements ClientInterface
@@ -33,7 +27,7 @@ test('DigitalOceanDropletCreator uploads a raw SSH key and returns droplet detai
         apiKey: 'test-token',
         name: 'sword-web',
         region: 'ams3',
-        size: 's-1vcpu-1gb',
+        serverType: 's-1vcpu-1gb',
         image: 'ubuntu-24-04-x64',
         publicKey: 'ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIB-example uploaded-key',
         publicIpPollAttempts: 1,
@@ -93,7 +87,7 @@ test('DigitalOceanDropletCreator resolves an existing key by name', function () 
         apiKey: 'test-token',
         name: 'sword-app',
         region: 'nyc1',
-        size: 's-1vcpu-2gb',
+        serverType: 's-1vcpu-2gb',
         image: 'ubuntu-24-04-x64',
         publicKey: 'existing-key',
         publicIpPollAttempts: 1,
